- cleanup and renaming GenericRequestFlows - adjusting per batch expectation in SearchJobDefinitions
