+++ src/main/scala/de/awagen/kolibri/base/actors/flows/GenericFlows.scala
@@ -1,0 +1,47 @@
+/**
+  * Copyright 2021 Andreas Wagenmann
+  *
+  * Licensed under the Apache License, Version 2.0 (the "License");
+  * you may not use this file except in compliance with the License.
+  * You may obtain a copy of the License at
+  *
+  * http://www.apache.org/licenses/LICENSE-2.0
+  *
+  * Unless required by applicable law or agreed to in writing, software
+  * distributed under the License is distributed on an "AS IS" BASIS,
+  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  * See the License for the specific language governing permissions and
+  * limitations under the License.
+  */
+
+package de.awagen.kolibri.base.actors.flows
+
+import akka.actor.{ActorRef, ActorSystem}
+import akka.http.scaladsl.Http
+import akka.http.scaladsl.model.{HttpRequest, HttpResponse}
+import akka.stream.scaladsl.Flow
+import de.awagen.kolibri.base.actors.tracking.ThroughputActor.AddForStage
+
+import scala.util.Try
+
+/**
+  * Flows to be used generically on RequestContext
+  */
+object GenericFlows {
+
+  case class Host(hostname: String, port: Int)
+
+  def getHttpConnectionPoolFlow[T](implicit actorSystem: ActorSystem): Host => Flow[(HttpRequest, T), (Try[HttpResponse], T), Http.HostConnectionPool] =
+    x => Http().cachedHostConnectionPool[T](x.hostname, x.port)
+
+  def getHttpsConnectionPoolFlow[T](implicit actorSystem: ActorSystem): Host => Flow[(HttpRequest, T), (Try[HttpResponse], T), Http.HostConnectionPool] =
+    x => Http().cachedHostConnectionPoolHttps[T](x.hostname, x.port)
+
+  def flowThroughActorMeter[U](meter: ActorRef, stageName: String): Flow[U, U, _] = {
+    Flow.fromFunction(x => {
+      meter ! AddForStage(stageName)
+      x
+    })
+  }
+
+}
